Remove doubles.  initDegutil() in degutil.c causes the game to crash when Quit is selected if converted to floats, so leaving this one alone for now.  The remaining doubles can't be removed until libBulletML has doubles removed.

+++ trunk/noiz2sa/src/barragemanager.h
@@ -23,7 +23,7 @@
 
 typedef struct {
   BulletMLParser *bulletml;
-  double maxRank, rank;
+  float maxRank, rank;
   int type;
   int frq;
 } Barrage;
